Move util package
Moves the util package to the webinterface package.
