+++ src/test/java/nl/tudelft/contextproject/webinterface/util/ActionUtilTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.util.webinterface;
+package nl.tudelft.contextproject.webinterface.util;
 
 import static org.mockito.Mockito.*;
